==> code/1.data_processing.ipynb <==
import numpy as np
import pandas as pd

from scipy.stats import chi2_contingency,ttest_ind
from scipy import stats
from sklearn.model_selection import train_test_split, KFold
from sklearn.metrics import auc, roc_curve
import matplotlib.pyplot as plt
import matplotlib
matplotlib.rcParams['font.sans-serif'] = ['FangSong']
matplotlib.rcParams['axes.unicode_minus'] = False

import statsmodels.api as sm
from sklearn.linear_model import LogisticRegression
from sklearn import svm
from sklearn import ensemble

data = pd.read_csv('C:/Users/Ca1/Desktop/papers/Nocturnal_hypertension by C/data.csv', encoding = 'gbk', index_col = 0)

dataplot = data[['Male','Age,y','BMI,kg/m2','Smoker','Alcohol intake','Diabetes mellitus','Hypertension','Hyperlipidemia'
,'Hypertension family history','CVD history','Antihypertension drugs','RAAS drugs intake','nRAAS drugs intake'
,'Proteinuria,mg/24h','Hemoglobin,g/L','BUN,mmol/L','Scr,mmol/L','Uric acid,mmol/L','FBG,mmol/L','Triglyceride,mmol/L'
,'Cholesterol,mmol/L','HDL-C,mmol/L','LDL-C,mmol/L','Serum albumin,g/L','Serum sodium,mmol/L','Serum calcium,mmol/L'
,'Serum phosphate,mmol/L','β2-MG,ug/mL','iPTH,pg/mL','Homocysteine,umol/L','eGFR,mL/ (min·1.73 m2)'
,'Clinic SBP,mmHg','Clinic DBP,mmHg','nocturnal_hypertension']]

dummy_col = [col for col in dataplot.columns if len(dataplot[col].unique()) < 5]
float_col = [col for col in dataplot.columns if col not in dummy_col]

# for col in float_col:
#     f,ax=plt.subplots(1,2,figsize=(12,5)) 
#     ax[0].boxplot(x=dataplot.loc[dataplot[col].isnull() == False,col].values)
#     ax[0].grid()
#     bins = np.linspace(min(dataplot[col].values),max(dataplot[col].values),20)
#     ax[1].hist(dataplot[col].values,bins)
#     ax[1].grid()
#     plt.xlabel(col,fontsize=15)
#     fig = plt.gcf() 
#     plt.show()

dummy_col = [col for col in dataplot.columns if len(dataplot[col].unique()) < 5]
float_col = [col for col in dataplot.columns if col not in dummy_col]
for col in float_col:
    data.loc[data[col]>data[col].mean()+3*data[col].std(),col] = np.nan
data=data.fillna(data.mean())
# data[float_col].describe()

dataplot = data[['Male','Age,y','BMI,kg/m2','Smoker','Alcohol intake','Diabetes mellitus','Hypertension','Hyperlipidemia'
,'Hypertension family history','CVD history','Antihypertension drugs','RAAS drugs intake','nRAAS drugs intake'
,'Proteinuria,mg/24h','Hemoglobin,g/L','BUN,mmol/L','Scr,mmol/L','Uric acid,mmol/L','FBG,mmol/L','Triglyceride,mmol/L'
,'Cholesterol,mmol/L','HDL-C,mmol/L','LDL-C,mmol/L','Serum albumin,g/L','Serum sodium,mmol/L','Serum calcium,mmol/L'
,'Serum phosphate,mmol/L','β2-MG,ug/mL','iPTH,pg/mL','Homocysteine,umol/L','eGFR,mL/ (min·1.73 m2)'
,'Clinic SBP,mmHg','Clinic DBP,mmHg','nocturnal_hypertension']]

# dataplot.info()
# dummy_col = [col for col in dataplot.columns if len(dataplot[col].unique()) < 5]
# float_col = [col for col in dataplot.columns if col not in dummy_col]
# for col in float_col:
#     f,ax=plt.subplots(1,2,figsize=(12,5)) 
#     ax[0].boxplot(x=dataplot.loc[dataplot[col].isnull() == False,col].values)
#     ax[0].grid()
#     bins = np.linspace(min(dataplot[col].values),max(dataplot[col].values),20)
#     ax[1].hist(dataplot[col].values,bins)
#     ax[1].grid()
#     plt.xlabel(col,fontsize=15)
#     fig = plt.gcf()
#     plt.show()

data.to_csv('C:/Users/Ca1/Desktop/papers/Nocturnal_hypertension by C//data_continuous_gbk_20201211.csv', encoding= 'gbk', index ='False')
